explicitly add theiaeuk to SE in preparation for ONT integration

+++ tasks/quality_control/comparisons/task_screen.wdl
@@ -72,7 +72,7 @@ task check_reads {
     # check total number of basepairs 
     bp_total=$(expr $read1_bp + $read2_bp)
     if [ "${bp_total}" -le "~{min_basepairs}" ]; then
-      fail_log+="; the total number of basepairs is below the minimum of ~{min_basepairs}"
+      fail_log+="; ${bp_total} read basepairs is below the minimum of ~{min_basepairs}"
     fi
 
     #checks four and five: estimated genome length and coverage
@@ -231,7 +231,7 @@ task check_reads_se {
     #checks four and five: estimated genome length and coverage
     if [ "~{skip_mash}" == "false" ]; then
       # estimate genome length if theiaprok AND expected_genome_length was not provided
-      if [ "~{workflow_series}" == "theiaprok" ] && [[ -z "~{expected_genome_length}" ]]; then
+      if ( [ "~{workflow_series}" == "theiaprok" ] || [ "~{workflow_series}" == "theiaeuk" ] ) && [[ -z "~{expected_genome_length}" ]]; then
         # First Pass; assuming average depth
         mash sketch -o test -k 31 -m 3 -r ~{read1} > mash-output.txt 2>&1
         grep "Estimated genome size:" mash-output.txt | \
@@ -247,28 +247,30 @@ task check_reads_se {
         estimated_coverage=`head -n1 coverage_output`
 
         # Check if second pass is needed
-        if [ ${estimated_genome_length} -gt "~{max_genome_length}" ] || [ ${estimated_genome_length} -lt "~{min_genome_length}" ] ; then
-          # Probably high coverage, try increasing number of kmer copies to 10
-          M="-m 10"
-          if [ ${estimated_genome_length} -lt "~{min_genome_length}" ]; then
-            # Probably low coverage, try decreasing the number of kmer copies to 1
-            M="-m 1"
+        if [ "${workflow_series}" == "theiaprok" ]; then
+          if [ ${estimated_genome_length} -gt "~{max_genome_length}" ] || [ ${estimated_genome_length} -lt "~{min_genome_length}" ] ; then
+            # Probably high coverage, try increasing number of kmer copies to 10
+            M="-m 10"
+            if [ ${estimated_genome_length} -lt "~{min_genome_length}" ]; then
+              # Probably low coverage, try decreasing the number of kmer copies to 1
+              M="-m 1"
+            fi
+    
+            mash sketch -o test -k 31 ${M} -r ~{read1} > mash-output.txt 2>&1
+            grep "Estimated genome size:" mash-output.txt | \
+              awk '{if($4){printf("%d", $4)}} END {if (!NR) print "0"}' > genome_length_output
+            grep "Estimated coverage:" mash-output.txt | \
+              awk '{if($3){printf("%d", $3)}} END {if (!NR) print "0"}' > coverage_output
+              
+            # remove mash outputs
+            rm -rf test.msh
+            rm -rf mash-output.txt
           fi
-
-          mash sketch -o test -k 31 ${M} -r ~{read1} > mash-output.txt 2>&1
-          grep "Estimated genome size:" mash-output.txt | \
-            awk '{if($4){printf("%d", $4)}} END {if (!NR) print "0"}' > genome_length_output
-          grep "Estimated coverage:" mash-output.txt | \
-            awk '{if($3){printf("%d", $3)}} END {if (!NR) print "0"}' > coverage_output
-          
-          # remove mash outputs
-          rm -rf test.msh
-          rm -rf mash-output.txt
+            
+          estimated_genome_length=`head -n1 genome_length_output`
+          estimated_coverage=`head -n1 coverage_output`
         fi
-        
-        estimated_genome_length=`head -n1 genome_length_output`
-        estimated_coverage=`head -n1 coverage_output`
-
+    
       # estimate coverage if theiacov OR expected_genome_length was provided
       elif [ "~{workflow_series}" == "theiacov" ] || [ "~{expected_genome_length}" ]; then
         if [ "~{expected_genome_length}" ]; then
